sane import pruning, refactor to reuse primitives for rebalance
